+++ CHANGELOG.md
@@ -67,3 +67,4 @@
 # 2.1.0 (7/08/2016-)
 * Added interaction methods with the child nodes to allow users to change the color of the node and the stroke color.
 * Added a way to query objects based on sub objects.
+* Lazily loaded the node interaction methods.