+++ CHANGELOG.md
@@ -68,3 +68,4 @@
 * Added interaction methods with the child nodes to allow users to change the color of the node and the stroke color.
 * Added a way to query objects based on sub objects.
 * Lazily loaded the node interaction methods.
+* Made graph backwards compatible with d3 v3.